Don't worry, I merged it in for you humans!

+++ lib/ExpediaBookings/src/main/java/com/expedia/bookings/data/payment/PointsAndCurrency.kt
@@ -2,4 +2,4 @@ package com.expedia.bookings.data.payment
 
 import com.expedia.bookings.data.Money
 
-data class PointsAndCurrency(var points: Int, var pointsType: PointsType, var amount: Money)
+data class PointsAndCurrency(var points: Float, var pointsType: PointsType, var amount: Money)
